@@ -1,46 +1,58 @@
 {
-  "name": "chasquifx",
-  "version": "1.0.0",
+  "name": "mimo-magic-starter",
   "private": true,
+  "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "build": "vite build && next build",
+    "start": "tsx scripts/dev.ts",
+    "dev": "tsx scripts/dev.ts",
+    "dev:frontend": "vite",
+    "dev:api": "next dev --port 3001"
   },
   "dependencies": {
-    "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.0",
-    "@mui/icons-material": "^7.1.0",
-    "@mui/material": "^7.1.0",
-    "@mui/x-date-pickers": "^8.3.1",
-    "@supabase/supabase-js": "^2.39.7",
-    "axios": "^1.9.0",
-    "cors": "^2.8.5",
-    "dayjs": "^1.11.13",
-    "dotenv": "^16.0.0",
-    "express-rate-limit": "^7.5.0",
-    "helmet": "^7.1.0",
-    "jsonwebtoken": "^9.0.2",
+    "@radix-ui/react-checkbox": "1.1.3",
+    "@radix-ui/react-dialog": "1.1.5",
+    "@radix-ui/react-label": "2.1.1",
+    "@radix-ui/react-popover": "1.1.5",
+    "@radix-ui/react-progress": "1.1.1",
+    "@radix-ui/react-select": "2.1.5",
+    "@radix-ui/react-slider": "1.2.2",
+    "@radix-ui/react-slot": "1.1.1",
+    "@radix-ui/react-switch": "1.1.2",
+    "@radix-ui/react-toggle": "1.1.1",
+    "@radix-ui/react-tooltip": "1.1.7",
+    "better-sqlite3": "^11.10.0",
+    "class-variance-authority": "0.7.1",
+    "clsx": "2.1.1",
+    "cmdk": "1.1.1",
+    "date-fns": "^3.6.0",
+    "dotenv": "16.4.7",
+    "esbuild": "0.25.1",
+    "express": "5.1.0",
+    "kysely": "^0.28.2",
+    "lucide-react": "0.474.0",
     "mongodb": "^6.0.0",
-    "mongoose": "^8.15.0",
-    "morgan": "^1.10.0",
-    "next": "^14.0.0",
-    "nodemailer": "^6.9.9",
-    "python-bridge": "^1.1.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^7.6.0",
-    "uuid": "^11.1.0",
-    "web-vitals": "^5.0.1",
-    "winston": "^3.11.0"
+    "next": "^14.2.29",
+    "react": "18.2.0",
+    "react-day-picker": "8.10.1",
+    "react-dom": "18.2.0",
+    "react-router-dom": "^7.6.1",
+    "tailwind-merge": "3.2.0",
+    "tailwindcss-animate": "1.0.7"
   },
   "devDependencies": {
-    "@types/node": "^20.0.0",
-    "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.0",
-    "eslint": "^8.0.0",
-    "eslint-config-next": "^14.0.0",
-    "typescript": "5.8.3"
+    "@types/express": "5.0.0",
+    "@types/node": "22.13.5",
+    "@types/react": "19.0.10",
+    "@types/react-dom": "19.0.4",
+    "@vitejs/plugin-react": "4.3.4",
+    "autoprefixer": "10.4.18",
+    "ignore": "7.0.3",
+    "postcss": "8.4.35",
+    "tailwindcss": "3.4.17",
+    "tsx": "4.19.3",
+    "typescript": "5.8.2",
+    "vite": "6.3.1"
   }
 }
